subset example uses non-projected map
For some reason, I get an error with check() in RStudio, when this is
there. The error message is odd, because it complains that the rgdal
package is not installed, but it *is* installed, and things work fine
interactively.

I fiddled with things in DESCRIPTION (suggests?  depends?) but decided
not to waste more time on it, since I think the problem may be that my R
setup is wonky. I was using Rdevel and then switched back to R-3.6.3 to
try to fix another probem, and I think the missing-but-not-missing rgdal
problem might be related to that.

Anyway, the example is a limited-area one, and so there's no actual need
for a map projection; plot,coastline-method() is suitable.