fiddling with dependencies...
I was hoping to get a cleaner message telling the builder what to do if
ncdf4 is not installed, but in the end I gave up, after an hour in the
quagmire of moving things around in DESCRIPTION, using importFrom tags,
and so forth. I've gone through this hassle several times for oce, and
really still do not understand the right recipe for a clean build.

In the end, I think I got back to where I was before I started fiddling.
I'm committing this just so I don't slide back, but I'm not finished
yet.

Next, I want to see if the build is clean whether ncdf4 is installed or
not.