'check' works with or without ncdf4
Although a warning is issues if ncdf4 is not available.  I don't see a
way around this, but I may have just got myself overly confused, with
moving ncdf4 between Required, Suggests, etc in DESCRIPTION, while also
adjusting importFrom and so forth.  Really, I should just Require it,
but the problem is that then R CHECK says we aren't using it (because
it's used by oce, not by argoFloats itself).

I may return to this later, but at least now the error that cropped up
recently is gone, so I think this is a good time to commit.