plot(...,which=map) can show bathymetry
I had hoped to make it draw bathymetry by default, but that causes an
error in the building of the vignette during 'R CMD check'.  I spent a
lot of time trying to figure out why, and gave up.  I think it may be a
problem with timing out in the server request (perhaps in interactive
usage, a download parameter is set up differently than it is in the
vignette-building stage.)

What I ended up doing was to put comments in example code, hinting as to
how to get the bathymetry plot.

Frankly, I don't think there is any point in trying to look deeper into
the problem.  For one thing, RStudio froze on me sometimes, and other
times crashed.  Perhaps there is something wrong in
marmap::getNOAA.bathy().