appveyor tag removed from README
I decided that getting this to work was just too arduous, with lots of
tests and fiddling being required, with the extra step of having to go
through the 'argoFloats' user, who is not me.  Besides, I finally got
travisCI working on its new server, and that accomplishes the goal of
displaying the results of an independent, automated checking.