trial method for handleFlags() but...
I am committing this to record the trial attempt, but it does not
coexist well with the oce version.  Actually, I *think* I can make it
coexist well but only by importing from oce, and that adds about 2M to
the built size of th package, and I just cannot see that cost as
commensurate with the benefit.

Given this, I am going to make simple function, as opposed to a
specialization of the oce generic.  I need to think about a good name.
So far, the following (ranked from 1 for the most favourable) are
bouncing around in my head:

1. useFlags()       -- PRO: simple
                       CON: a bit vague
2. applyFlags()     -- PRO: simple
                       CON: people might think it means setting (like applying paint)
3. setFlaggedToNA() -- PRO: clear for default case.
                       CON: we only set to NA in default case
4. obeyFlags()      -- PRO: simple
                       CON: 'obey' is see useFlags
5. removeFlagged()  -- PRO:
                       CON: we do not actually *remove*, in any case