remove app.R because it blocks package build
I know what needs to be changed, but the work will take me a while, so I
am removing this app.R file, to permit the whole package to build.

I will likely try to put it back into the package within a week or two.
The argoFloats co-developers know how to access this app through a
different (non-public) route, so nothing is lost by my action today.