FINALLY figured out why [[ failed noninteractively
Jeeze Louise.  Why was this so mysterious?  It was not dispatching to
the oce function `[[,argo-method` because I had not done
requireNamespace("oce") first.  Why this worked in RStudio, I don't
know. Perhaps it loads or caches namespaces in some way that standalone
R does not. In any case, I don't really care.  The code 11_spiciness.R
in our paper repo runs fine both interactively and non-interatively, and
that seems like a fine test.