changes during JH/DK z
1. fix a problem in subset by cycle
2. rebuild indices (need to avoid failures in build tests)

There are > 10 errors now because of the changes to the indices, and JH
will look at them.  A lot are tests that do not contribute very much,
e.g. testing for precise names on the server (broken because we now
start with ifremer, not usgodae), testing the number of items in objects
(which will change as indices are rebuilt ... we should keep a few
tests, but certainly not repeated ones).