index dates decoded with lubridate::fast_strptime()
Previously, I used as.POSIXct() for this, but fast_strptime() saves
about 1s of executation time on my machine (macbook pro 2020; 2GHz
quadcore i5; 16 GB 3733 MHz LPDDR4X).  Admittedly, that gain will not be
noticeable to users, but there seems to be little harm in switching,
since my tests verify that the return values of the two functions match
on the core-argo index. I am retaining the timing information as
comments in the code, along with my test suite.  This will make it
trivial to return to the old scheme if necessary.

See https://github.com/ArgoCanada/argoFloats/issues/334 for discussion.

I instrumented the code to try a series of methods, timing the
computations and ensuring that the results matched.  The code
```R
library(argoFloats)
index <- getIndex(age=0)
```
was run with Rscript, and the test was performed 3 times, with results
as follows. Note that the switch ought to give a real-world speedup of 1
second.

as.POSIXct() timing:
   user  system elapsed
  2.841   0.048   2.913
  2.793   0.051   2.867
  2.850   0.041   2.899
strptime() timing:
   user  system elapsed
  2.435   0.008   2.451
  2.426   0.002   2.429
  2.440   0.002   2.444
ymd_hms() timing:
   user  system elapsed
  3.383   0.101   3.496
  3.340   0.102   3.445
  3.399   0.095   3.499
lubridate::fast_strptime() timing:
   user  system elapsed
  1.894   0.003   1.903
  1.860   0.002   1.865
  1.913   0.002   1.918
readr::parse_datetime() timing:
Warning: 1667 parsing failures.
  row col               expected        actual
 8956  -- date like %Y%m%d%H%M%S 2.0040404e+13
 8967  -- date like %Y%m%d%H%M%S 2.0040723e+13
 9025  -- date like %Y%m%d%H%M%S 2.0040423e+13
13984  -- date like %Y%m%d%H%M%S 2.0041001e+13
14024  -- date like %Y%m%d%H%M%S 2.0040325e+13
..... ... ...................... .............
See problems(...) for more details.

   user  system elapsed
  3.324   0.083   3.475
  3.293   0.080   3.433
  3.067   0.080   3.182