try to make webpage work
The autobuild was failing wrt bathy files that were downloaded.  And we
cannot have downloads for CRAN work, anyway. I am not sure whether to
use \dontrun or \donttest, or whether either will ensure a clean CRAN
build. Or perhaps we'll need to comment-out the bathy-related examples
in the docs; at some point, the benefit of testing the examples may not
be worth the hassle of finding a way to get the package to build
properly in the CRAN context.