remove use of marmap in examples
I'm checking this in to see if pkgdown (run by a github action) will work properly this way.  If it does, I'll then remove the internal use of marmap (by commenting it out, in case we need it later, if the topo-server API changes and marmap handles that change before oce handles it).