subset(index, cycle=<NUMERIC>) improved
I ran into a problem with a case where there were two cycles, one with
name ending in "D" or something like that ... actually, I don't remember
the details, but it was something simpple that caused a stop() and I
fixed it quickly while in the middle of something else.

Also, improved some debugging statements.