remove some topo downloads BUT
In RStudio, the 'check' operation complains about the existence of a
topo file.  So either an example or a test must be downloading such a
file.  I think that means that a plot() call is causing that to happen.
I added bathymetry=FALSE to a few cases but I am still seeing such a
file.  We cannot have that for a CRAN submission.

improve subset processingLog entry; remove some topo