fix broken URLs:
1. ifremer
2. usgodae
3. several youtube links ... not all, but I changed all, just in case

My procedure is to change \url{} to \code{} every time we get a stated
problem during an `R CMD CHECK` or a remote check. This might seem
Draconian, but I don't see any advantage in the docs having links that
will fail, even if occassionally.  After all, it's very easy for users
to copy the text and paste it into a web-browser.