solve dblclick problem of single-float view
The problem was that we were relying on secondary effects.  We were
updating the UI widget that holds float ID. So, we got a plot, then the
ID update got noticed and we got a second plot.  (And, later, we got an
identical third plot.)

My solution was to avoid this scheme of updating the UI widget.  Yes,
that's elegant in a programming way (it avoids repeating some lines of
code) but if the cost is an annoying redraw and a confusion in the undo
scheme, it ain't worth it.

So now, in this commit, we don't get the momentary single-cycle plot
when a double click is done.  We still get a flash redraw of the
single-float graph.  I'll look at that problem later.  It's not terribly
annoying, though; it just looks like a momentary flutter of the graph.
And it is not wrecking the undo scheme. Therefore I won't fret a whole
lot about it.