remove a UI update that seems to have no effect
I thought maybe this would eliminate the double-drawing of the map, but
it does not.  Still, removing the update seems to have no visible effect
on mapApp(), and so I commented out the UI update, on the grounds that
irrelevant code ought to be skipped over.  (I am not 100% sure on
whether there might be effects, so I only commented-out the code,
instead of, for now, deleting it.)