Address issues 584, 585 and 587, + related things
This is an aggregating commit for changes done in an attempt to address
the issue of the long time it can take `getIndex()` to work (which
caused time-outs in some systems of shared work).  I think the addition
of the ability to save and later read an original (.gz) file may help
a bit in speed, but my guess is that the other aspects of actually
reading that .gz file will still be problematic.  For that reason, the
best plan for folks will likely be to copy a saved .rda file with a new
name so it won't get rewritten later.  This is all explained in the
docs, which should be consulted in preference to these notes.

CR, JH and DK had a z meeting on Mar 1 to discuss the motivation for
this work, and JH and I had another on Mar 2 to talk about progress.
Given this, I won't get into too much detail here, preferring instead to
show a simple list^[In John Carmack's format with * meaning done and +
meaning not-yet-done.].

Note that getIndex() saves an RDA file by default, after downloading and
then discarding a gz file.  But if KEEP is set to TRUE, the gz file is
stored also.

TO DO LIST:

* 1. Ensure that KEEP=TRUE works.
* 2. Pair cached items with their creation times.
* 3. Ensure that the 'age' parameter is obeyed not just for files, but for cached values
   also.  I think this matters most for age=0 cases.
* 4. Permit filename to be a local .gz file (if server=NULL).
* 5. Permit filename to be a local .rda file (if server=NULL).
* 6. Write a 'white paper' (sandbox/dek/17_get_local_file) to
  demonstrate these things.
* 7. Fix issue 588 so I can build the white paper just mentioned.
* 8. Add an appendix to the main vignette about working with
  already-downloaded index files.

Items 1 and 2 are mainly for development work, but they may be
quite handy for that.

Items 3 and 4 are the main addition that will affect users.

Item 6 is for discussion between developers.